test: u2f device
