+++ package.json
@@ -23,13 +23,14 @@
   "homepage": "https://github.com/bcoin-org/bledger#readme",
   "scripts": {
     "browserify": "browserify -s bledger lib/bledger.js | uglifyjs -c > build/bledger.js",
-    "clean": "rm -f build/bledger.js build/vendor*",
+    "clean": "rm -f build/bledger.js build/vendor* .config/certs/*",
     "lint": "eslint lib/ test/ || exit 0",
     "test": "mocha --reporter spec test/*-test.js",
-    "test-device": "mocha --reporter spec test/device-hid/*-test.js",
-    "webpack": "webpack --config webpack/config.js",
-    "webpack-dev": "webpack --config webpack/config.dev.js",
-    "webpack-devdeps": "webpack --config webpack/config.devdeps.js"
+    "test-hid": "mocha --reporter spec test/device/hid-test.js",
+    "test-u2f": "karma start .config/karma.conf.js",
+    "webpack": "webpack --config .config/webpack.js",
+    "webpack-dev": "webpack --config .config/webpack.dev.js",
+    "webpack-devdeps": "webpack --config .config/webpack.devdeps.js"
   },
   "dependencies": {
     "bcoin": "^1.0.0-beta.15",
@@ -46,9 +47,14 @@
     "babelify": "^8.0.0",
     "browserify": "^14.5.0",
     "eslint": "^4.9.0",
+    "karma": "^2.0.0",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-mocha": "^1.3.0",
+    "karma-mocha-reporter": "^2.2.5",
+    "karma-webpack": "^2.0.9",
     "mocha": "^4.0.1",
-    "uglifyjs-webpack-plugin": "^1.0.0-beta.3",
     "uglify-es": "^3.1.3",
+    "uglifyjs-webpack-plugin": "^1.0.0-beta.3",
     "webpack": "^3.8.1"
   },
   "engines": {
